Lazy load CodeEditor from a separate chunk/bundle
