+++ src/components/inputs/CodeEditor.tsx
@@ -1,112 +1,11 @@
-import React, {useState, useEffect, useCallback} from 'react';
-import {Box} from '@mui/material';
+// src/components/CodeEditor.tsx
+import React, {Suspense} from 'react';
+import {asyncDecorator} from '@plotly/dash-component-plugins';
+import codeeditor from '../../utils/LazyLoader/CodeEditor';
+import type {ModuleDefinition, CompletionItem} from '../../fragments/CodeEditorAutocomplete';
 import {DashComponentProps} from 'props';
-import CodeMirror from '@uiw/react-codemirror';
-import {python} from '@codemirror/lang-python';
-import {vscodeLight, vscodeDark} from '@uiw/codemirror-theme-vscode';
-import {
-  AutocompleteManager,
-  ModuleDefinition,
-  CompletionItem
-} from '../../fragments/CodeEditorAutocomplete';
-import {indentUnit} from '@codemirror/language';
-import {EditorState} from '@codemirror/state';
 
-/**
- * Code Editor component with configurable Python module autocompletion.
- */
-const CodeEditor = ({
-  id = 'code-editor',
-  value = '',
-  height = '400px',
-  width = '100%',
-  margin = 2,
-  readOnly = false,
-  lineNumbers = true,
-  tabSize = 2,
-  highlightActiveLine = true,
-  darkTheme = false,
-  moduleDefinitions = {},
-  setProps
-}: CodeEditorProps) => {
-  const [editorLoaded, setEditorLoaded] = useState(false);
-
-  // Check if we're in the browser environment
-  useEffect(() => {
-    if (typeof window !== 'undefined') {
-      setEditorLoaded(true);
-    }
-  }, []);
-
-  // Handle changes
-  const handleChange = useCallback(
-    (newValue: string) => {
-      if (setProps) {
-        setProps({value: newValue});
-      }
-    },
-    [setProps]
-  );
-
-  // Create autocomplete manager
-  const autocompleteManager = new AutocompleteManager(moduleDefinitions);
-
-  // Configure extensions with custom completions
-  const extensions = [
-    python(),
-    autocompleteManager.createCompletionExtension(),
-    indentUnit.of(' '.repeat(tabSize)), // Set tab size
-    EditorState.tabSize.of(tabSize) // Set tab size for tab character
-  ];
-
-  return (
-    <Box id={id} m={margin} width={width} height={height}>
-      {editorLoaded ? (
-        <CodeMirror
-          value={value}
-          height={height}
-          width={width}
-          theme={darkTheme ? vscodeDark : vscodeLight}
-          extensions={extensions}
-          onChange={handleChange}
-          readOnly={readOnly}
-          basicSetup={{
-            lineNumbers,
-            highlightActiveLineGutter: true,
-            highlightSpecialChars: true,
-            foldGutter: true,
-            dropCursor: true,
-            allowMultipleSelections: true,
-            indentOnInput: true,
-            syntaxHighlighting: true,
-            bracketMatching: true,
-            closeBrackets: true,
-            autocompletion: true,
-            rectangularSelection: true,
-            crosshairCursor: true,
-            highlightActiveLine,
-            highlightSelectionMatches: true,
-            closeBracketsKeymap: true,
-            defaultKeymap: true,
-            searchKeymap: true,
-            historyKeymap: true,
-            foldKeymap: true,
-            completionKeymap: true,
-            lintKeymap: true
-          }}
-        />
-      ) : (
-        // Box fallback
-        <Box bgcolor={darkTheme ? '#1E1E1E' : '#FFFFFF'} p={1}>
-          {value || 'Loading editor...'}
-        </Box>
-      )}
-    </Box>
-  );
-};
-
-// Add the moduleDefinitions type to our props
-type CodeEditorProps = {
+type CodeEditorPropsType = {
   /** The initial value of the editor */
   value?: string;
   /** The height of the editor */
@@ -129,6 +28,54 @@ type CodeEditorProps = {
   moduleDefinitions?: ModuleDefinition;
 } & DashComponentProps;
 
-// Re-export types from Autocomplete for external use
-export type {ModuleDefinition, CompletionItem};
+// Main component (gets exported)
+const CodeEditor = ({
+  id = 'code-editor',
+  value = '',
+  height = '400px',
+  width = '100%',
+  margin = 2,
+  readOnly = false,
+  lineNumbers = true,
+  tabSize = 2,
+  highlightActiveLine = true,
+  darkTheme = false,
+  moduleDefinitions = {},
+  setProps
+}: CodeEditorPropsType) => {
+  return (
+    <ControlledCodeEditor
+      id={id}
+      value={value}
+      height={height}
+      width={width}
+      margin={margin}
+      readOnly={readOnly}
+      lineNumbers={lineNumbers}
+      tabSize={tabSize}
+      highlightActiveLine={highlightActiveLine}
+      darkTheme={darkTheme}
+      moduleDefinitions={moduleDefinitions}
+      setProps={setProps}
+    />
+  );
+};
+
+// Async-decorated version (internal)
+const RealCodeEditor = asyncDecorator(CodeEditor, () =>
+  Promise.all([codeeditor()]).then(([codeeditor]) => codeeditor)
+);
+
+// Controlled version with Suspense (internal)
+const ControlledCodeEditor: React.FC<CodeEditorPropsType> = (props) => {
+  return (
+    <Suspense fallback={<div>Loading code editor...</div>}>
+      <RealCodeEditor {...props} />
+    </Suspense>
+  );
+};
+
+// Export the original component AND the types AND the default props
 export default CodeEditor;
+export type {CodeEditorPropsType};
+export type {ModuleDefinition, CompletionItem};
