+++ src/utils/LazyLoader/CodeEditor.js
@@ -1,0 +1,1 @@
+export default () => import(/* webpackChunkName: "codeeditor" */ '../../fragments/CodeEditor');
